sharding: begin shard local storage
