+++ sharding/interfaces.go
@@ -1,6 +1,7 @@
 package sharding
 
 import (
+	"github.com/ethereum/go-ethereum/common"
 	cli "gopkg.in/urfave/cli.v1"
 )
 
@@ -22,3 +23,12 @@ type Service interface {
 // ServiceConstructor defines the callback passed in when registering a service
 // to a sharding node.
 type ServiceConstructor func(ctx *cli.Context) (Service, error)
+
+// ShardBackend defines an interface for a shardDB's necessary method
+// signatures.
+type ShardBackend interface {
+	Get(k common.Hash) (*[]byte, error)
+	Has(k common.Hash) bool
+	Put(k common.Hash, val []byte) error
+	Delete(k common.Hash) error
+}
